set theme before app renders to the DOM
