+++ src/pages/_app.tsx
@@ -1,17 +1,20 @@
 import { AppProps } from 'next/app';
 import Head from 'next/head';
 import Layout from '../components/layout';
+import { ThemeProvider } from '../context/Theme';
 import '../styles/index.css';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <Layout>
+    <ThemeProvider>
       <Head>
         <title>FrontendTony</title>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Component {...pageProps} />
-    </Layout>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </ThemeProvider>
   );
 }
