+++ src/types/index.ts
@@ -1,0 +1,10 @@
+export interface ThemeObject {
+  [key: string]: string;
+}
+
+export type Theme = 'light' | 'dark' | null;
+
+export interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
